Display budget info even if no time has been tracked to that task.
