+++ src/@types/global.d.ts
@@ -16,19 +16,4 @@ declare global {
   };
 
   type Trello = TrelloSync & TrelloAsync;
-
-  type HarvestAPIResponse = {
-    per_page: number;
-    total_pages: number;
-    total_entries: number;
-    next_page: number | null;
-    previous_page: number | null;
-    page: number;
-    links: {
-      first: string;
-      next: string | null;
-      previous: string | null;
-      last: string;
-    };
-  };
 }
